==> notebooks/1-similarityModelling/otherExperiments/3.0-bl-analyzeCrossAblationOutput.ipynb <==
import pandas as pd
import pickle 
import numpy as np

modelList = ["finalModel", "noLayers", "noRDrop", "noHeadTail", "noTranslated"]

for modelSuffix in modelList: 
    print("MODEL: " + str(modelSuffix))
    inPath = "/home/blitt/projects/localNews/models/sentEmbeddings/3.0-crossModelAblation/" + modelSuffix
    with open(inPath + "/outputData.pkl", "rb") as f: 
        metrics = pickle.load(f)

    with open(inPath + "/time.pkl", "rb") as f: 
        time = pickle.load(f)
    
    print("RUNNING TIME: ")
    print(time)
    
    EPOCHS = 3
    FOLDS = 5

    iterList = []
    corrList = []
    #go through each epoch 
    for epoch in range(EPOCHS): 
        corrList = []
        for fold in range(FOLDS):

            df = pd.DataFrame(metrics[fold][epoch]).T
            df.columns =  ["loss", "pred", "true"]
            corr = np.corrcoef(df["pred"], df["true"])[1,0]
            corrList.append(corr)
        print("Epoch: " + str(epoch))
        print(corrList)
        print("Standard Deviation: " + str(np.std(corrList)))
        print("Average Correlation: " + str(np.mean(corrList)))
        print("---------------------")
    print("\n\n")
