==> notebooks/1-similarityModelling/otherExperiments/4.1-bl-ReadCrossEncoderOnTest.ipynb <==
import os
import pandas as pd
import numpy as np

#just need to inspect the file that was written with results from the final model on the test set
randomSeeds = [135, 200, 60, 85, 92]

baseDir = "/shared/3/projects/newsDiffusion/models/3.0-crossModelAblation/finalOnTest/"

for seed in randomSeeds: 
    currDf = pd.read_csv(baseDir + str(seed) + "testDf.tsv", sep="\t")
    np.corrcoef(currDf["ground_truth"], currDf["scaledSims"])[0, 1]

currDf
